Allows transfer of objects in GUI

+++ src/particle/particle/Particle.cpp
@@ -2,7 +2,7 @@
 
 #include "Application.h"
 
-Particle::Particle(int x, int y, float size, Vec2f dir, uint16_t lifetime, glm::vec4 colour)
+Particle::Particle(float x, float y, float size, Vec2f dir, uint16_t lifetime, glm::vec4 colour)
 	: x(x), y(y), m_Size(size), m_Dir(dir), m_Lifetime(lifetime), m_Age(0), m_Colour(colour)
 {
 }
